+++ src/app.ts
@@ -1,0 +1,10 @@
+import React from "react";
+
+const button = React.createElement('button',
+  {
+    onClick: () => {
+      alert('clicked');
+    }
+  }, 'button');
+
+export default button;